requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-admin-kubi==1.1.0
django-allauth==64.2.1
django-allauth-themes==0.1.2
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
